Move subq reading into plaform-specific source files with a common header
