+++ src/meson.build
@@ -64,6 +64,12 @@ sources = [
                       fallback: 'release')
 ]
 
+if host_machine.system() == 'darwin'
+    sources += 'subq_read_macos.c'
+else
+    sources += 'subq_read_mmc.c'
+endif
+
 # Optional, for printing MusicBrainz submission URLs as QR codes
 qrencode = dependency('libqrencode', version: '>= 4.0', required: false,
                       static: static_build)
